extract utils mod, StoreError
